Cleanup COM Wrapping
